backend/requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.1.2
Flask-CORS==4.0.0
requests==2.31.0
lxml==4.9.3
pandas>=2.1.1
numpy>=1.26.0
openpyxl==3.1.2
fpdf2==2.8.7
cssselect==1.2.0
playwright==1.41.2
celery==5.3.6
redis==5.0.2
